+++ app/src/main/kotlin/com/ediposouza/teslesgendstracker/interactor/FirebaseParsers.kt
@@ -50,7 +50,7 @@ class DeckParser(
         val patch: String = "",
         val views: Int = 0,
         val likes: List<String> = listOf(),
-        val cards: Map<String, Int> = mapOf(),
+        val cards: Map<String, Long> = mapOf(),
         val updates: Map<String, Map<String, Int>> = mapOf(),
         val comments: Map<String, Map<String, String>> = mapOf()
 ) {
